backend/requirements.txt
fastapi==0.115.5
uvicorn==0.30.6
pydantic==2.9.2
python-multipart==0.0.9

google-genai==0.3.0

Pillow==10.4.0
python-dotenv==1.0.1
replicate==0.32.1
